Handling orders of resolved responses of multiple async calls
1. Need to make 3 api calls: getPositions, getInstruments, getQuotes (in order)
2. getInstruments is dependent on the response of getPositions and getQuotes is dependent on the response of getInstruments
3. In order to achieve 'synchronous' api calls, I used async/await.

4. getPositions returns an array of my stock positions. Loop around this array and extract the instrument urls and pass that information to dispatch(getInstruments(url)). Store each dispatched result into a new array called getInstrumentFetches. This new array will have elements of unresolved promises.

Issue) I used Promise.all(getInstrumentFetches) to resolve promises hoping to preserve the order of promises passed in. However, it seems like the order of resolved promises is dependent on which promises get resolved first. (I am not hundred percent sure). Therefore, I had to keep track of the index of each promise in 'getQuotesFetches' and sort 'quotes' based on those indices.